requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
celery==5.3.4
redis==5.0.1
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.12.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
boto3==1.34.0
pymupdf==1.23.7
pypdf2==3.0.1
ghostscript==0.7
pytesseract==0.3.10
pillow==10.1.0
python-magic==0.4.27
clamd==1.0.2
aioredis==2.0.1
httpx==0.25.2
python-pptx==0.6.23
pydantic-settings==2.1.0
aiofiles==23.2.1